add paste rules for links
