+++ packages/frontend/@n8n/design-system/src/components/N8nMarkdownEditor/composables/useMarkdownEditor.ts
@@ -1,0 +1,123 @@
+import type { EditorView } from '@tiptap/pm/view';
+/** These probably should be lazy loaded */
+import Link from '@tiptap/extension-link';
+import Placeholder from '@tiptap/extension-placeholder';
+import Strike from '@tiptap/extension-strike';
+import TableCell from '@tiptap/extension-table-cell';
+import TableHeader from '@tiptap/extension-table-header';
+import TableRow from '@tiptap/extension-table-row';
+import TaskItem from '@tiptap/extension-task-item';
+import TaskList from '@tiptap/extension-task-list';
+import { Table } from '@tiptap/extension-table';
+import { Markdown } from '@tiptap/markdown';
+import StarterKit from '@tiptap/starter-kit';
+import { useEditor } from '@tiptap/vue-3';
+import { watch } from 'vue';
+import type { SetupContext } from 'vue';
+
+import { MarkdownSlashCommandExtension, renderSlashCommandMenu } from '../extensions';
+import type { N8nMarkdownEditorEmits, N8nMarkdownEditorProps } from '../MarkdownEditor.types';
+import {
+	copyMarkdown,
+	getEditorContent,
+	pasteMarkdown,
+	setEditorContent,
+} from '../markdownEditorUtils';
+
+export const useMarkdownEditor = (
+	props: Readonly<N8nMarkdownEditorProps>,
+	emit: SetupContext<N8nMarkdownEditorEmits>['emit'],
+) => {
+	const baseExtensions = [
+		StarterKit,
+		Strike,
+		Link.configure({
+			openOnClick: false,
+		}),
+		Placeholder.configure({
+			placeholder: () => props.placeholder ?? '',
+			showOnlyCurrent: false,
+		}),
+		Table.configure({
+			resizable: true,
+		}),
+		TableRow,
+		TableHeader,
+		TableCell,
+		TaskList,
+		TaskItem.configure({
+			nested: true,
+		}),
+		Markdown.configure({
+			markedOptions: {
+				gfm: true,
+				breaks: false,
+			},
+		}),
+		MarkdownSlashCommandExtension.configure({
+			render: renderSlashCommandMenu,
+		}),
+	];
+
+	const editor = useEditor({
+		content: getEditorContent(props.modelValue),
+		contentType: props.modelValue ? 'markdown' : undefined,
+		extensions: [...baseExtensions, ...(props.extensions ?? [])],
+		editable: !props.disabled && !props.readonly,
+		editorProps: {
+			...props.editorProps,
+			attributes: {
+				class: 'n8n-markdown',
+				'data-test-id': 'n8n-markdown-editor-content',
+				...props.editorProps?.attributes,
+			},
+			handleDOMEvents: {
+				...props.editorProps?.handleDOMEvents,
+				copy: (view: EditorView, event: ClipboardEvent) =>
+					props.editorProps?.handleDOMEvents?.copy?.(view, event) ||
+					(editor.value ? copyMarkdown(editor.value, event) : false),
+				paste: (view: EditorView, event: ClipboardEvent) =>
+					props.editorProps?.handleDOMEvents?.paste?.(view, event) ||
+					(editor.value ? pasteMarkdown(editor.value, event) : false),
+			},
+		},
+		onCreate: ({ editor }) => {
+			emit('ready', editor);
+		},
+		onUpdate: ({ editor }) => {
+			const markdown = editor.getMarkdown();
+
+			emit('update:modelValue', markdown);
+			emit('input', markdown);
+			emit('change', markdown);
+		},
+		onFocus: ({ event }) => {
+			emit('focus', event);
+		},
+		onBlur: ({ editor, event }) => {
+			emit('blur', editor.getMarkdown(), event);
+		},
+	});
+
+	watch(
+		() => props.modelValue,
+		(value) => {
+			if (!editor.value) return;
+
+			const currentMarkdown = editor.value.getMarkdown();
+
+			if (currentMarkdown === value) return;
+
+			setEditorContent(editor.value, value ?? '');
+		},
+	);
+
+	watch(
+		() => [props.disabled, props.readonly] as const,
+		([disabled, readonly]) => {
+			editor.value?.setEditable(!disabled && !readonly);
+		},
+	);
+
+	return editor;
+};
