+++ packages/frontend/@n8n/design-system/src/locale/lang/en.ts
@@ -125,11 +125,6 @@ export default {
 	'markdownEditor.heading1': 'Heading 1',
 	'markdownEditor.heading2': 'Heading 2',
 	'markdownEditor.heading3': 'Heading 3',
-	'markdownEditor.link': 'Link',
-	'markdownEditor.linkUrlLabel': 'Link URL',
-	'markdownEditor.linkUrlPlaceholder': 'https://example.com',
-	'markdownEditor.applyLink': 'Apply link',
-	'markdownEditor.removeLink': 'Remove link',
 	'markdownEditor.slashCommandMenuLabel': 'Markdown commands',
 	'aiAssistant.textMessage.restoreVersion': 'Restore version',
 	'aiAssistant.versionCard.restoreModal.title': 'Restore to this version?',
